requirements.txt
Flask
langchain  # Si la estás usando
google-generativeai